registration done up to the where it should get registerSuccess, other exceptions work I shall fix the null pointer issue and should have registration finished asap tomorrow-loginIncrement

+++ VSS-web/src/main/java/com/team33/entities/dao/AccountDao.java
@@ -9,7 +9,7 @@ public interface AccountDao {
 
 	  public List<Account> getAccounts() throws DataAccessException;
 
-	  public Account getAccount(Long accountId) throws DataAccessException;
+	  public Account getAccount(Integer accountId) throws DataAccessException;
 
 	  public Account getAccount(String username) throws  DataAccessException;
