+++ VSS-web/src/main/java/com/team33/services/AccountService.java
@@ -17,7 +17,7 @@ import org.springframework.dao.DataAccessException;
 public interface AccountService {
     public List<Account> getAccounts() throws DataAccessException;
 
-	  public Account getAccount(Long accountId) throws DataAccessException;
+	  public Account getAccount(Integer accountId) throws DataAccessException;
 
 	 
 	  public void registerAccount(Account account) throws DataAccessException;
